+++ crates/miden-agglayer/src/eth_types/mod.rs
@@ -1,9 +1,12 @@
-pub mod address;
+pub mod eth_account_id_format;
+pub mod eth_address_format;
+
 pub mod amount;
 pub mod global_index;
 pub mod metadata_hash;
 
-pub use address::EthAddressFormat;
 pub use amount::{EthAmount, EthAmountError};
+pub use eth_account_id_format::EthAccountIdFormat;
+pub use eth_address_format::{AddressConversionError, EthAddressFormat};
 pub use global_index::{GlobalIndex, GlobalIndexError};
 pub use metadata_hash::MetadataHash;